requirements.txt
pyrogram>=1.0.0
tgcrypto
python-dotenv
